M1555 Refactoring for TS
