+++ src/style/NewAttributeModal.module.scss
@@ -1,0 +1,18 @@
+@use './theme.scss';
+
+tr td:first-child {
+  background: theme.$tableRowEven;
+}
+
+.details-table {
+  border: solid 1px theme.$secondary;
+}
+
+.styled-row {
+  justify-content: space-between;
+  gap: 1rem;
+}
+
+.input-container {
+  width: 100%;
+}
