+++ src/style/theme.scss
@@ -2,6 +2,10 @@
 $primary: #174b82;
 $secondary: #dddce4;
 
+$tableRowOdd: #fdfdfe;
+$tableRowEven: #f8f8fa;
+$tableRowHover: #f1f1f4;
+
 /** Media Query Break Points **/
 $mobile-break: 599px;
 $tablet-portrait-break: 600px;
