+++ tsconfig.json
@@ -3,6 +3,7 @@
     "target": "ESNext",
     "lib": ["dom", "dom.iterable", "esnext"],
     "allowJs": true,
+    "allowImportingTsExtensions": false,
     "checkJs": false,
     "skipLibCheck": true,
     "esModuleInterop": true,
